fix(basic-select): переименовать компонент, удалить комментарии, пофиксить тест, изменить ФС
