+++ src/components/BasicSelect/__tests__/Container.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { render, RenderResult } from '@testing-library/react';
 
-import { Container } from '../components/Container/Container';
+import { Container } from '../Container';
 
 const renderComponent = (props): RenderResult => {
   const { ...restProps } = props;
