--- src/components/Select/BasicSelect/index.ts
@@ -1,1 +1,0 @@
-export { BasicSelect } from './BasicSelect';
